% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simon_pr.R
\name{as_flextable.simon_pr}
\alias{as_flextable.simon_pr}
\title{Convert \linkS4class{ph2simon4} to \link[flextable]{flextable}}
\usage{
\method{as_flextable}{simon_pr}(x, ...)
}
\arguments{
\item{x}{a \linkS4class{ph2simon4}}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \code{\link[=as_flextable.simon_pr]{as_flextable.simon_pr()}} returns a \link[flextable]{flextable}
}
\description{
Convert \linkS4class{ph2simon4} to \link[flextable]{flextable}
}
\keyword{internal}
