% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{gg_ph2simon}
\alias{gg_ph2simon}
\alias{autoplot.ph2simon}
\alias{autoplot.ph2simon4}
\alias{autolayer.ph2simon4}
\alias{autolayer.ph2simon}
\title{Plot Simon's Two-Stage Design}
\usage{
\method{autoplot}{ph2simon}(object, ...)

\method{autoplot}{ph2simon4}(object, ...)

\method{autolayer}{ph2simon4}(
  object,
  r1 = object@r1,
  n1 = object@n1,
  r = object@r,
  n = object@n,
  pu = object@pu,
  pa = object@pa,
  ...
)

\method{autolayer}{ph2simon}(object, ...)
}
\arguments{
\item{object}{a \link[clinfun]{ph2simon} or \linkS4class{ph2simon4} object}

\item{...}{parameters of function \code{\link[=ph2simon4]{ph2simon4()}}, most importantly \code{type}}

\item{r1, n1, r, n}{(optional) \link[base]{integer} scalars, see \linkS4class{ph2simon4}.}

\item{pu, pa}{\link[base]{double} scalars, see function \link[clinfun]{ph2simon}}
}
\value{
Function \code{\link[=autoplot.ph2simon]{autoplot.ph2simon()}} returns a \link[ggplot2]{ggplot} object.

Function \code{\link[=autolayer.ph2simon]{autolayer.ph2simon()}} returns a \link[base]{list} of \link[ggplot2]{ggproto} and labels.
}
\description{
Plot \link[clinfun]{ph2simon} object using \CRANpkg{ggplot2}.
}
\keyword{internal}
