% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power_ph2simon}
\alias{power_ph2simon}
\alias{power.ph2simon4}
\alias{power.ph2simon}
\alias{powerCurve.ph2simon4}
\alias{powerCurve.ph2simon}
\title{Power of \link[clinfun]{ph2simon} or \linkS4class{ph2simon4} Object}
\usage{
power.ph2simon4(
  prob,
  object,
  ...,
  r1 = object@r1,
  n1 = object@n1,
  r = object@r,
  n = object@n
)

power.ph2simon(prob, object, ...)

powerCurve.ph2simon4(object, ...)

powerCurve.ph2simon(object, ...)
}
\arguments{
\item{prob}{\link[base]{double} scalar, true response rate \eqn{p}}

\item{object}{a \link[clinfun]{ph2simon} or \linkS4class{ph2simon4} object}

\item{...}{parameters of function \code{\link[=ph2simon4]{ph2simon4()}}, most importantly \code{type}}

\item{r1, n1, r, n}{(optional) \link[base]{integer} scalars, see \linkS4class{ph2simon4}.}
}
\description{
Power of \link[clinfun]{ph2simon} or \linkS4class{ph2simon4} Object
}
