% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2simon_etc.R
\name{print.ph2simon}
\alias{print.ph2simon}
\title{Print \link[clinfun]{ph2simon} Object}
\usage{
\method{print}{ph2simon}(x, ...)
}
\arguments{
\item{x}{\link[clinfun]{ph2simon} object}

\item{...}{potential parameters, currently not in use}
}
\value{
\link{print.ph2simon} does not have a return value
}
\description{
Print \link[clinfun]{ph2simon} object, overwriting \code{clinfun:::print.ph2simon}
}
\examples{
library(clinfun)
(x = ph2simon(pu = .2, pa = .4, ep1 = .05, ep2 = .1)) 
class(x)
autoplot(x, type = 'minimax')
autoplot(x, type = 'optimal')
autoplot(x, type = 'n1')
autoplot(x, type = 'maximax')

}
