% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph2simon_etc.R
\name{print_ph2simon}
\alias{print_ph2simon}
\title{Alternate Print Method for a Simon's Two-Stage Design}
\usage{
print_ph2simon(x, ...)
}
\arguments{
\item{x}{a \link[clinfun]{ph2simon} object}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{print_ph2simon} does not have a returned value.
}
\description{
An alternate \link[base]{print} method for \link[clinfun]{ph2simon} object.
}
\note{
We do not overwrite \code{clinfun:::print.ph2simon}.
}
\examples{
library(clinfun)
(x = ph2simon(pu = .2, pa = .4, ep1 = .05, ep2 = .1)) 
print_ph2simon(x)

}
