% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simon_pr.R
\name{show,Simon_pr-method}
\alias{show,Simon_pr-method}
\title{Show \linkS4class{Simon_pr} Object}
\usage{
\S4method{show}{Simon_pr}(object)
}
\arguments{
\item{object}{\linkS4class{Simon_pr} object}
}
\value{
The \link[methods]{show} method for \linkS4class{Simon_pr} object
does not have a returned value.
}
\description{
Show \linkS4class{Simon_pr} object
}
