% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show_simon}
\alias{show_simon}
\alias{show,simon_oc-method}
\alias{show,simon_pr-method}
\alias{show,ph2simon4-method}
\title{Show S4 Objects}
\usage{
\S4method{show}{simon_oc}(object)

\S4method{show}{simon_pr}(object)

\S4method{show}{ph2simon4}(object)
}
\arguments{
\item{object}{S4 objects, e.g., \linkS4class{simon_oc}, \linkS4class{simon_pr}}
}
\value{
The \link[methods]{show} method for S4 objects
does not have a returned value.
}
\description{
Show S4 object.
}
\keyword{internal}
