% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibs.R
\name{simon_bib}
\alias{simon_bib}
\alias{.simon89}
\title{\link[utils]{bibentry} of \CRANpkg{VisualizeSimon2Stage}}
\usage{
.simon89(key = "Simon89", ...)
}
\arguments{
\item{key, ...}{additional parameters of function \link[utils]{bibentry}}
}
\description{
\link[utils]{bibentry} of \CRANpkg{VisualizeSimon2Stage}
}
\keyword{internal}
