% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simon_pr.R
\docType{class}
\name{simon_design-class}
\alias{simon_design-class}
\title{S4 Class \linkS4class{simon_design}}
\description{
S4 Class \linkS4class{simon_design}
}
\section{Slots}{

\describe{
\item{\code{n1,n}}{\link[base]{integer} scalars, stage-1 sample size \eqn{n_1} and total sample size \eqn{n}}

\item{\code{r1,r}}{\link[base]{integer} scalars, number of positive response
in stage-1 \eqn{r_1} and overall \eqn{r} required \strong{exclusive}.
In other words, passing Stage 1 indicates observing \eqn{>r_1} positive responses,
and rejecting \eqn{H_0} indicates observing \eqn{>r} positive responses.}
}}

