% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simon_oc.R
\docType{class}
\name{Simon_oc-class}
\alias{Simon_oc-class}
\title{S4 class \linkS4class{Simon_oc}}
\description{
S4 class \linkS4class{Simon_oc}
}
\section{Slots}{

\describe{
\item{\code{prob}}{\strong{named} \link[base]{numeric} vector of length \eqn{p}, true response rate(s)}

\item{\code{maxResp}}{\link[base]{integer} vector of length \eqn{p},
the frequencies of each regime having maximum response.
The summation of \code{maxResp} is the number of simulation copies.}

\item{\code{Simon_maxResp}}{\link[base]{integer} vector of length \eqn{p},
the frequencies of each regime having maximum response and success in Simon's two-stage trial.}

\item{\code{eN}}{\link[base]{numeric} vector of length \eqn{p},
expected sample sizes by simulation}
}}

