% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simon_pr.R
\docType{class}
\name{Simon_pr-class}
\alias{Simon_pr-class}
\title{S4 class \linkS4class{Simon_pr}}
\description{
S4 class \linkS4class{Simon_pr}
}
\section{Slots}{

\describe{
\item{\code{.Data}}{\eqn{p\times 3} \link[base]{numeric} \link[base]{matrix}, probability of frail
(i.e., early termination), fail (to reject the null) and success (to reject the null), at each
true response rate given in \code{@prob}}

\item{\code{eN}}{\link[base]{numeric} vector of length \eqn{p}, expected sample size(s)}

\item{\code{prob}}{\link[base]{numeric} vector of length \eqn{p}, true response rate(s)}
}}

