% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simon_oc.R
\name{Sprintf.Simon_oc}
\alias{Sprintf.Simon_oc}
\title{Short Paragraph to Describe a \linkS4class{Simon_oc} Object}
\usage{
Sprintf.Simon_oc(model, ...)
}
\arguments{
\item{model}{\linkS4class{Simon_oc} object}

\item{...}{additional parameters, currently not in use}
}
\value{
\link{Sprintf.Simon_oc} returns a \link[base]{noquote} \link[base]{character} scalar.
}
\description{
To create a short paragraph to describe a \linkS4class{Simon_oc} object.
}
