% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sprintf.R
\name{Sprintf_ph2simon}
\alias{Sprintf_ph2simon}
\alias{Sprintf.ph2simon4}
\alias{Sprintf.ph2simon}
\title{Short Paragraph to Describe a \link[clinfun]{ph2simon} Object}
\usage{
Sprintf.ph2simon4(model, ...)

Sprintf.ph2simon(model, ...)
}
\arguments{
\item{model}{a \link[clinfun]{ph2simon} or \linkS4class{ph2simon4} object}

\item{...}{parameters of function \code{\link[=ph2simon4]{ph2simon4()}}, most importantly \code{type}}
}
\value{
Function \code{\link[=Sprintf.ph2simon]{Sprintf.ph2simon()}} and \code{\link[=Sprintf.ph2simon4]{Sprintf.ph2simon4()}} return a \link[base]{character} scalar.
}
\description{
To create a short paragraph to describe a \link[clinfun]{ph2simon} object.
}
\keyword{internal}
