% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{VisualizeSimon2Stage-package}
\alias{VisualizeSimon2Stage}
\alias{VisualizeSimon2Stage-package}
\title{VisualizeSimon2Stage: Visualize Simon's Two-Stage Design}
\description{
To visualize the probabilities of early termination, fail and success of Simon's two-stage design. To evaluate and visualize the operating characteristics of Simon's two-stage design.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tingtingzhan/VisualizeSimon2Stage}
}

}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com}

Authors:
\itemize{
  \item Tingting Zhan \email{tingtingzhan@gmail.com}
}

}
