% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0S4.R, R/ph2simon4.R
\docType{class}
\name{ph2simon4}
\alias{ph2simon4}
\alias{ph2simon4-class}
\title{S4 Object \linkS4class{ph2simon4}}
\usage{
ph2simon4(object, type = "minimax", ...)
}
\arguments{
\item{object}{a \link[clinfun]{ph2simon} object}

\item{type}{\link[base]{character} scalar or \link[base]{vector}, see \strong{Slots}}

\item{...}{additional parameters, currently not in use}
}
\description{
One or more Simon's two-stage designs.
}
\section{Slots}{

\describe{
\item{\code{r1,r}}{non-negative \link[base]{integer} scalars or \link[base]{vector}s}

\item{\code{n1,n}}{positive \link[base]{integer} scalars or \link[base]{vector}s}

\item{\code{pu,pa}}{\link[base]{double} scalars}

\item{\code{alpha,beta}}{\link[base]{double} scalars}

\item{\code{nmax}}{\link[base]{integer}}

\item{\code{type}}{\link[base]{character} scalars or \link[base]{vector}s,
type of Simon's two-stage design,}
}}

\keyword{internal}
