% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_ph2simon}
\alias{print_ph2simon}
\title{Alternate Print Method for a Simon's Two-Stage Design}
\usage{
print_ph2simon(x, ...)
}
\arguments{
\item{x}{a \link[clinfun]{ph2simon} object}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \code{\link[=print_ph2simon]{print_ph2simon()}} does not have a returned value.
}
\description{
An alternate \link[base]{print} method for \link[clinfun]{ph2simon} object.
}
\note{
We do not overwrite function \code{clinfun:::print.ph2simon}.
}
\keyword{internal}
