% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simon_oc.R
\docType{class}
\name{Simon_oc-class}
\alias{Simon_oc-class}
\title{S4 Class \linkS4class{Simon_oc}}
\description{
Operating characteristics of Simon's two-stage design.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{\linkS4class{Simon_pr} object}

\item{\code{maxResp}}{\link[base]{integer} \link[base]{vector} of length \eqn{p},
the frequencies of each regime having maximum response.
The summation of \code{maxResp} is the number of simulation copies.}

\item{\code{Simon_maxResp}}{\link[base]{integer} \link[base]{vector} of length \eqn{p},
the frequencies of each regime having maximum response and success in Simon's two-stage trial.}
}}

