% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simon_pr.R
\docType{class}
\name{Simon_pr-class}
\alias{Simon_pr-class}
\title{S4 Class \linkS4class{Simon_pr}}
\description{
Probabilities of early termination, failure and success, of Simon's Two-Stage Design.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{\eqn{l\times 3} \link[base]{numeric} \link[base]{matrix}, probability of frail
(i.e., early termination), fail (to reject the null) and success (to reject the null), at each
response rate \eqn{p} given in \verb{@prob}}

\item{\code{eN}}{\link[base]{numeric} \link[base]{vector} of length \eqn{l}, expected sample size(s) \eqn{\textrm{E}(N)}}

\item{\code{prob}}{\link[base]{numeric} \link[base]{vector} of length \eqn{l}, response rate(s) \eqn{p}}
}}

