% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sprintf.R
\name{Sprintf.simon_oc}
\alias{Sprintf.simon_oc}
\title{Short Paragraph to Describe a \linkS4class{simon_oc} Object}
\usage{
Sprintf.simon_oc(model, ...)
}
\arguments{
\item{model}{\linkS4class{simon_oc} object}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \code{\link[=Sprintf.simon_oc]{Sprintf.simon_oc()}} returns a \link[base]{character} scalar.
}
\description{
To create a short paragraph to describe a \linkS4class{simon_oc} object.
}
\keyword{internal}
