% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeSimon2Stage_info.R
\docType{package}
\name{VisualizeSimon2Stage-package}
\alias{VisualizeSimon2Stage-package}
\title{Visualize Simon\'s Two-Stage Design}
\description{
Functions for visualizing the probabilities of early termination, fail and success of
Simon's two-stage design.
Functions for evaluating and visualizing the operating characteristics of
Simon's two-stage design.
}
\keyword{package}
