% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{VisualizeSimon2Stage-package}
\alias{VisualizeSimon2Stage}
\alias{VisualizeSimon2Stage-package}
\title{VisualizeSimon2Stage: Visualize Simon's Two-Stage Design}
\description{
To visualize the probabilities of early termination, fail and success of Simon's two-stage design. To evaluate and visualize the operating characteristics of Simon's two-stage design.
}
\note{
Help files of individual functions are intentionally suppressed in the \code{pdf} manual.
Users are encouraged to get started with

\code{vignette('intro', package = 'VisualizeSimon2Stage')}
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

Authors:
\itemize{
  \item Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})
}

}
