% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{VisualizeSimon2Stage-package}
\alias{VisualizeSimon2Stage}
\alias{VisualizeSimon2Stage-package}
\title{VisualizeSimon2Stage: Visualize Simon's Two-Stage Design}
\description{
To visualize the probabilities of early termination, fail and success of Simon's two-stage design. To evaluate and visualize the operating characteristics of Simon's two-stage design.
}
\references{
\doi{10.1016/0197-2456(89)90015-9}

\url{https://www.ncss.com/software/pass/}
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

}
